edit make.bat

+++ os.001/make.bat
@@ -1,4 +1,8 @@
-del /Q os.dat
+@echo off
+if exist os.dat del /Q os.dat
+@echo on
 nasm os.asm -f bin -o os.dat -l os.dat.lst -DBUILDBOOT
-del /Q os.dsk
+@echo off
+if exist os.dsk del /Q os.dsk
+@echo on
 nasm os.asm -f bin -o os.dsk -l os.dsk.lst -DBUILDDISK
